requirements.txt
beautifulsoup4==4.12.2
camelot_py==0.11.0
contractions==0.1.73
numpy==1.25.1
pandas==2.0.3
pikepdf==7.2.0
python_docx==0.8.11
python_magic==0.4.27
rdflib==7.0.0
regex==2023.6.3
tika==1.24
xmltodict==0.13.0
flag-embedding
torch
transformers